+++ examples/hello_world/config.py
@@ -8,8 +8,8 @@ import numpy as np
 # Expose the enviroment to python
 import neworder
 
-# Timeline is compulsory - define a dummy timeline
-neworder.timespan = np.array([0, 1])
+# Timeline is compulsory, and must be float type - define a dummy timeline
+neworder.timespan = np.array([0, 1], dtype=float)
 # We only need one timestep
 neworder.timestep = neworder.timespan[1]
